+++ Sources/PortavozCore/Identifiers.swift
@@ -20,3 +20,9 @@ public struct SummaryID: Hashable, Codable, Sendable, RawRepresentable {
     public init(rawValue: UUID) { self.rawValue = rawValue }
     public init() { self.rawValue = UUID() }
 }
+
+public struct AudioAssetID: Hashable, Codable, Sendable, RawRepresentable {
+    public let rawValue: UUID
+    public init(rawValue: UUID) { self.rawValue = rawValue }
+    public init() { self.rawValue = UUID() }
+}